web/base_django/requirements.txt
Django>=3.0,<4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-cors-middleware==1.5.0
django-bootstrap4==1.1.1
django-tables2==2.2.1
django-filter==2.2.0
django-markup==1.4.1
docutils==0.16
whitenoise==5.0.1
gunicorn==20.0.4
PyYAML==5.3
requests==2.23.0
django-graphql-jwt==0.3.0
graphene-django==2.8.2
pytest-django==3.8.0
psycopg2-binary>=2.8,<3.0
django-import-export==1.2.0
pyufbr
django_countries
pillow
django-grappelli==2.14.1
drf-yasg








